Add files via upload
The project involves the scraping of homeopathic medicine data from the online medicine delivery platform 1mg using Python library Beautiful soup. The collected data was analyzed to provide insights and recommendations for a consultancy firm's client who wants to open a homeopathic medicine store. The data was used to estimate the total cost of opening the store, and an interactive dashboard was created to help the client select the best medicine for their store using Python libraries, Pandas, NumPy, Seaborn, SQL, Excel and Power BI.